+++ apps/client/src/app/pages/about/terms-of-service/terms-of-service-page.routes.ts
@@ -3,12 +3,12 @@ import { publicRoutes } from '@ghostfolio/common/routes/routes';
 
 import { Routes } from '@angular/router';
 
-import { TermsOfServicePageComponent } from './terms-of-service-page.component';
+import { GfTermsOfServicePageComponent } from './terms-of-service-page.component';
 
 export const routes: Routes = [
   {
     canActivate: [AuthGuard],
-    component: TermsOfServicePageComponent,
+    component: GfTermsOfServicePageComponent,
     path: '',
     title: publicRoutes.about.subRoutes.termsOfService.title
   }