+++ apps/client/src/app/pages/portfolio/allocations/allocations-page.routes.ts
@@ -2,12 +2,12 @@ import { AuthGuard } from '@ghostfolio/client/core/auth.guard';
 
 import { Routes } from '@angular/router';
 
-import { AllocationsPageComponent } from './allocations-page.component';
+import { GfAllocationsPageComponent } from './allocations-page.component';
 
 export const routes: Routes = [
   {
     canActivate: [AuthGuard],
-    component: AllocationsPageComponent,
+    component: GfAllocationsPageComponent,
     path: '',
     title: $localize`Allocations`
   }